requirements.txt
streamlit
tiktoken
qdrant-client
python-dotenv